Rename file for easy import
PR feedback - renamed the navbar file to index.tsx for easier import. navbar.less kept the same since there's only one needed in index.tsx, and it makes it easier to understand what component its styling
