+++ src/App.tsx
@@ -10,7 +10,7 @@ import BlockTemplate from './containers/template-1-col-block/Template';
 import GridTemplate from './containers/template-24-col-grid/Template';
 
 import NotFound from './containers/not-found/NotFound';
-import NavBar from './components/navbar/NavBar';
+import NavBar from './components/navbar';
 import Footer from './components/Footer';
 import { Layout } from 'antd';
 const { Content } = Layout;
